requirements.txt
ollama==0.3.0
PyYAML==6.0.1
questionary==2.0.1
rich==13.7.1